Rearrange folders

+++ src/main/java/com/learnenglish/studentsvocabulary/infrastructure/controller/UserController.java
@@ -1,10 +1,10 @@
-package com.learnenglish.studentsvocabulary.controller;
+package com.learnenglish.studentsvocabulary.infrastructure.controller;
 
 import com.learnenglish.studentsvocabulary.dtos.CreateUserRequestDto;
 import com.learnenglish.studentsvocabulary.dtos.LoginUserResponseDto;
 import com.learnenglish.studentsvocabulary.model.User;
 import com.learnenglish.studentsvocabulary.model.Vocabulary;
-import com.learnenglish.studentsvocabulary.service.LogInService;
+import com.learnenglish.studentsvocabulary.infrastructure.authentication.LogInService;
 import com.learnenglish.studentsvocabulary.service.UserService;
 import com.learnenglish.studentsvocabulary.service.VocabularyService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,7 +12,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
-import java.util.Set;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.function.Predicate;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/users")
@@ -39,17 +43,28 @@ public class UserController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable int id){
+    public void deleteUser(@PathVariable int id, @RequestHeader(value="Authorization") String bearerToken){
+        if(!logInService.isLogedIn(id, bearerToken))
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Missing Authorization header");
+
         userService.delete(id);
     }
 
     @GetMapping("/{id}/vocabularies")
-    public Set<Vocabulary> vocabularies(@PathVariable int id,
-                                        @RequestHeader(value="Authorization") String bearerToken){
+    public List<Vocabulary> vocabularies(@PathVariable int id,
+                                        @RequestHeader(value="Authorization") String bearerToken,
+                                        @RequestParam Map<String,String> params){
         if(!logInService.isLogedIn(id, bearerToken))
             throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Missing Authorization header");
 
-        return userService.getVocabularies(id);
+        List<Predicate<Vocabulary>> vocabularyPredicateList = new ArrayList<>();
+        params.forEach((key,value)->{
+            if(key.equals("favorite") && value.equals("true")) vocabularyPredicateList.add(v-> v.isFavorite());
+            if(key.equals("category") && value.equals("phrasal-verb")) vocabularyPredicateList.add(v-> v.isPhrasalVerb());
+        });
+
+        Predicate<Vocabulary> intersectQuerys = vocabularyPredicateList.stream().reduce(v->true, Predicate::and);
+        return userService.getVocabularies(id).stream().filter(intersectQuerys).collect(Collectors.toList());
     }
 
     @PostMapping("/{id}/vocabularies")
